requirements.txt
azure-core==1.29.3
azure-search-documents==11.6.0b3
faiss-cpu==1.7.4
openai==0.28
numpy==1.24.2
pandas==2.2.2
pydantic==2.3.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
streamlit==1.32.2
tenacity==8.2.2
tiktoken==0.5.1
seaborn
matplotlib